updated add company form
