+++ components/MadeInYoungstown/Companies.tsx
@@ -1,8 +1,8 @@
-import { companies } from "../data/companies";
+import { companies } from "../../data/companies";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faLocationDot } from "@fortawesome/free-solid-svg-icons";
-import DataCard from "./Globals/DataCard";
-import Search from "./Globals/Search";
+import DataCard from "../Globals/DataCard";
+import Search from "../Globals/Search";
 import { useState } from "react";
 
 export default function Companies() {
